Add support for command-line arguments and options
Previously, the YAML and iCal directories needed to be specified in
const.py. The locations specified in const.py have been modified to
represent the default locations, if no command-line options are
provided to override them.

In addition, parameters have provided to execute the script in one of
two modes:

  * To execute the script in Jenkins test job mode, the -t/--test
    parameters can be used. Valid values include 'check', 'gate',
    and 'post'.

  * To execute the script in YAML-iCal conversion mode, the
    -c/--convert parameter can be used. This can be combined with
    the -m/--meetings parameter, to specify a list of meetings
    to process. The meeting list file should contain the filenames
    of meetings that correspond to meeting files in the YAML
    directory, and should be separated by newlines.

Support for command line options was implemented using the argparse
module. This module has been added to requirements.txt.

Change-Id: I9e5e8f123d7a28997cbd8eb8e67b246d68c97260