Add Search Team Meeting
Search is a new project being split out of Glance based on the
Catalog Index Service, which was developed and released in Kiloi[1].
At the Liberty Summit we decided to broaden the scope and pursue
making it its own project with a goal to provide advanced and scalable
search across multi-tenant cloud resources. The final name for the search
project is still being finalized.

[1] https://review.openstack.org/#/c/138051/

Change-Id: Iedbfd6fc0c1244458af0a53edeb9887d3c7427ff