Remove Install Guide Update Meeting
This meeting is in a Google Hangout, but this repository lists
IRC meetings.  The OpenStack project asks constituent projects
to use IRC as a primary forum for regular meetings so that they
are accessible to the whole community, participants can easily
join meetings casually as needed, and logs of meetings are
available for people unable to attend.

See http://governance.openstack.org/reference/irc.html for more
detail.

Change-Id: Ic02f514eb00b986bdff7067e2872db8ff6066332