Remove on hold meeting
The nova-net to neutron migration meeting has been on hold for
some time. At the summit, those interested in driving this
forward have changed from those who attended this meeting.

Best to remove this meeting from the schedule and allow the
new folks to decide what date/time works best for them, if
the choose to create a specific meeting.

Change-Id: I1aba5b40c7d60dc3cdbd98c79f3cb0acfba978c8