Move DefCore Meetings to openstack-meeting-4
After we got settled on an alternate-week schedule for DefCore meetings,
it was pointed out that we'll actually avoid a potential conflict with
other meetings in the official channels by holding our 1500 UTC meetings
on odd weeks instead of even weeks.  Since there's no conflict, we'd
prefer to host both meetings on the same official meeting channel.

Change-Id: Id1ea35bde76457beebbf023875c171a2993bd70f