Remove the product team meeting
The product working group has switched to using telephony for
their team meeting:

http://eavesdrop.openstack.org/meetings/product_working_group/2015/product_working_group.2015-07-06-19.01.log.html
http://eavesdrop.openstack.org/irclogs/%23openstack-meeting/%23openstack-meeting.2015-07-13.log.html#t2015-07-13T19:06:25

Since it is no longer an IRC meeting, it does not merit an entry
in this system.  We also ask that projects conduct their team
meetings in IRC so that people unable to attend the meeting are
able to read the logs, and people who are generally on IRC
might be able to be pinged from within meetings.  See
http://governance.openstack.org/reference/irc.html for more info.

Change-Id: I237b0ed5133b5b911c85838ae2aec45b74e696e5