Change DefCore Meeting Time
At the beginning of the current cycle, the DefCore Committee adopted
an alternating-week schedule on a trial basis in an attempt to foster
more global participation. [1]  However, upon evaluation following
the conclusion of the trial period, the Committee found that the
schedule was actually detrimental to participation and elected to
move back to a single meeting time. [2]  This patch updates the
meeting calendar accordingly.

[1] https://etherpad.openstack.org/p/DefCoreFlag.2
[2]
http://eavesdrop.openstack.org/meetings/defcore/2015/defcore.2015-07-15-15.00.log.html#l-77

Change-Id: I43d393218343d472634ecfab99c588b5e02f42a6