Erased i18n meeting_id for proper past irclog link
'I18N Team Meeting' information is well displayed
on eavesdrop.o.o. However, 'past meetings' references
to the wrong link like
: http://eavesdrop.openstack.org/meetings/i18nteam/ .
This is mainly because meeting_id is set to 'i18nteam'.
It can be changed to the correct name like
'OpenStack I18n Meeting' but spaces may cause problems.
So I would propose simply to delete this meeting_id
because i18n Agenda link actually contains previous
meeting logs and it is not essential to have meeting_id
on this file.

Change-Id: If5f2421b7c1259d581ec4eb95515bd26085624fb