neutron-upgrades: moved the meeting to -alt slot that is free now
We wanted to have the time slot booked weekly for the subteam, but did
not have it every week before. Now that it's free since
I6c5f5245ec17790dd1ef7253c984e97e01a1ee8c merged, we are happy to book
the channel for our needs.

Hopefully we still have it available. ;)

Change-Id: I880a7a0ea96be20b51df86d7d0e6b5cbaeb92444