Add explict start dates to meeting yaml
This is an intermediate step, but valuable, step in stabilising the IRC
meeting iCal.  There are a couple of issues.

1. Every time we regenerate the iCal we set a new start point.
2. The odd/even decision is based on the start date (which defaults to
   the time the ics file is generated)

By adding explicit start dates to all out meetings (which have been
reset this week) we mitigate both of these issues.

There are other aspects that require updates to yaml2ical, but this
relatively simple step while we get those items in place.

Note: this week you can verify that these start_dates are correct by
1. running tox on master
2. save output/irc-meetings.ical to output/master.ical
3. apply this patch and re-run tox
4. diff output/master.ical output/irc-meetings.ical

Change-Id: I7ae70c684884c0c708738392a0b43bbb3c1c878e