Use yaml2ical from git rather than pypi
There are a coupel of new features that have landed in yaml2ical recently.
It would be great to take advantage of them when publishing/generating our
calendar data.

This review pulls yaml2ical from git rather than pypi.  It is inteneded to be
reverted as soon as yaml2ical 0.6.0 is tagged and released.

Change-Id: I00a282e0624f185073f5b42f9613a9eb1e4e8a93