Correct times for skipped third-party meetings
In I81386e018bf66c167cf77fc23fea847a6fc39cb7 we added entries to skip
8 third party meetings.  However, due to this being the first use of this
I didn't notice that all 8 were associated with the "Tuesday" meeting
as opposed to 4 for each of Monday and Tuesday.

This generated a valid ICS file but 4 of the meetings had the wrong time.

Move 4 items from Monday to Tuesday.

Change-Id: Ia2c1f9faadd5ce664ea43da9c77cd651b7bcc53a