Make sorting consistent on smaller devices
Unfortunately using the batch_meetings() function causes the content
to become unsorted on smaller devices. This could be seen on phones or
if making the windows size small on a desktop browser.

Instead of using batch_meetings() use the column-count CSS property.
This allows the index list to switch from 1 to 4 columns based on the page
width

Change-Id: Iebae34488412cc35ca3c9f38a43897d30d20d8db