Switch to odd, update chairs
This IRC meeting happens every other week. It was mistakenly
set to the even weeks, this commit switches to odd weeks to
match the working groups schedule.

This commit also removes one chair (John) who is no longer
participating with this group and adds a chair (krotscheck)
who recently joined.

Change-Id: If1759c415ad707c643a26b6ba323b4462b375600