Proposing OpenStack SWG Meeting
This proposal is based on a doodle poll at:

http://doodle.com/poll/kdnr9kfgaamw4873

The odd weeks will be on Thursdays and the even weeks on Tuesdays. We
are currently in the 29th week, and I propose this week that we meet
on Thursday, and on Tuesday next week.

Change-Id: I32981becb63784e88a994cf77f20c7b9ed9dc881
Related: Ib1ff3b8e62183b7038becce34c140564c214e8f1