move HA meeting an hour later, to 0900 UTC
This reverts the switch to daylight savings introduced by
Ice9c8460c900af8db1ec6be51ef7133d5b5f39d4.  The main motivation is to
find a time which Andrew Beekhof (in Australia) can make it, without
introducing new problems for someone else.

http://eavesdrop.openstack.org/meetings/ha/2016/ha.2016-06-13-07.07.log.html#l-128
http://eavesdrop.openstack.org/irclogs/%23openstack-ha/%23openstack-ha.2016-06-20.log.html#t2016-06-20T08:00:49

Moving it a further hour later (1000 UTC) would be better for Andrew,
but would be too late for the NTT guys in Japan.  Those of us in Europe
seem to be OK with 0900 UTC.  We don't have any Americans yet, which is
lucky because that would make it impossible to find a good time for
everyone ;-)

Change-Id: Iaa547f07292a7369158694483498674ba6d02010