Add a UTC clock for ease of use
It's a bit difficult to use the meeting schedule page
if you're not in UTC time.

This patch adds a little clock at the top (in UTC time)
for those in far away time zones who don't want to keep track.

The clock uses JavaScript, and is set up properly
to display a message on unsupported browsers.

To test, clone this change, run `tox`, and serve from the
output/ directory (use `python -m SimpleHTTPServer PORT`).

Change-Id: I8ec40caa425199a2efef2a0bf345dfbb07054482