move HA meeting to Wednesday 0900 UTC
Monday 0900 UTC is a bit awkward for Europeans, since lots of stuff
happens at the beginning of the week, and it also makes it difficult
to prepare for the meeting.

So as discussed in last week's meeting, move it to the same time
on Wednesday:

http://eavesdrop.openstack.org/meetings/ha/2016/ha.2016-12-05-09.10.log.html#l-102

To avoid a channel conflict, also change to #openstack-meeting-alt.

Change-Id: Ic0911688520e13f10efa258cb60f8315ee5b97e8