Update doc team meeting and remove doc/user-guides meeting
The current time schedule for the documentation team meeting is
not working for the team. This is the next best option with the
remaining available slots.

The doc/user-guides team meeting has also been removed in this
patch. The specialty teams were disbanded within the documentation
team and therefore the lead does not require a specific time to
meet any longer.

Change-Id: If36dc5ebe1b935233e3cdd2cca1ecdd6528d68d6