Allow for meetings to be hosted in project channel
It was recently agreed[0] that project teams should be able to host IRC
meetings in their own channels. This patch removes the check forcing
meetings to happen in the existing meeting channels, allowing for
projects to specify their office hours as well as an alternative IRC
channel for their meetings.

[0] I631e2597bf818c5e4b2f813da14764d2cb55d0ce

Change-Id: Icd3557d46bbe6ee0ed94aa0dde3a056516bddd40