Propose a new time for weekly keystone meeting
On of the action items that came out of the keystone retrospective
during the Rocky PTG was that we should try and move our meeting to
be more accomodating to developers in APAC timezones. The consensus
was to move it two hours earlier.

In order to avoid a conflict, we'll have to move meeting rooms as
well.

Change-Id: Ic5d22e2804a51beecde8b5f35759a3d256239c70