Remove the policy meeting
This meeting was put in place about 18 months ago for the purpose of
getting a handle on the direction of RBAC and policy enforcement
across OpenStack. Since then we've put together community goals [0][1]
as part of a roadmap for improving policy across the platform.

Meetings on this topic have become redundant and rarely include new
topics since most involved parties understand the basic approach
behind what we're trying to accomplish. It's also apparent that we
feel pretty comfortable approaching (and being approached by) teams in
order to answer questions about policy and RBAC.

[0] https://governance.openstack.org/tc/goals/queens/policy-in-code.html
[1] http://specs.openstack.org/openstack/keystone-specs/specs/keystone/rocky/define-default-roles.html

Change-Id: Iebc2de0b630483709ec499b39e9418d17a2014e5