Remove Cinder-Nova cross-project meeting
As the coordinated multi-attach activities got finished the teams are keeping
touch on the project channels and not utilizing the cross-project meeting
slot actively anymore.

This patch removes the meeting from the list of OpenStack Meetings.

We will resurrect the meeting series in the future and add it back in case
it becomes necessary again.

Change-Id: Ifc029b38c34350df784dc1de48348c4f583e4787