Creating an IRC meeting for CloudKitty
This adds a yaml file to describe CloudKitty IRC meetings.

In an effort to reconnect with the community, the cloudkitty team wants to
re-schedule IRC meetings. These will be held on the first friday of each month
at 15h00 UTC. The first meeting will exceptionnaly be held on the second friday
of november, on the 09/11/2018.

Change-Id: Iab07b40808f24daa00b80986b9f3e2f86a45eb4c
Story: 2004180
Task: 27666