Change watcher meeting schedule and chair
Watcher's last meeting was in December 2018 around the time Alexander
resigned as PTL. I have discussed with contributors and the current PTL
to reintroduce the meetings. This information was discussed on the
discuss mailing list. It is not absolutely sure if licanwei wants to be
chair but I have no reason to believe not so.

Change-Id: Ie04c92618b9dccbf47fb26e587a67544504196f2